+++ source/BuildingBlocks.Application/BuildingBlocks.Application.csproj
@@ -9,6 +9,7 @@
     <ItemGroup>
       <PackageReference Include="AspNetCore.HealthChecks.AzureServiceBus" Version="8.0.0" />
       <PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="8.0.0" />
+      <PackageReference Include="AspNetCore.HealthChecks.AzureStorage" Version="7.0.0" />
       <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.17.2" />
       <PackageReference Include="Energinet.DataHub.Core.App.FunctionApp" Version="8.1.0" />
       <PackageReference Include="Microsoft.ApplicationInsights" Version="2.22.0" />
